refactor Literals out of org.whdl.intermediate.expressions and into org.whdl.intermediate
