+++ src/main/java/org/whdl/intermediate/ExprEvalVisitor.java
@@ -2,6 +2,8 @@ package org.whdl.intermediate;
 
 public class ExprEvalVisitor {
 
+  // FIXME this is backwards; copy the current ExprTypeVisitor
+  
   public Object visit(Expression e){
     return e.accept(this);
   }
