+++ src/test/java/org/whdl/intermediate/TestExprTypeVisitor.java
@@ -1,8 +1,8 @@
 package org.whdl.intermediate;
 
 import static org.junit.Assert.*;
+
 import org.junit.Test;
-import org.whdl.intermediate.expressions.*;
 import org.whdl.intermediate.types.*;
 
 public class TestExprTypeVisitor {
